Add files via upload
Write a function in JS fixage that receives an array of integers. The integers represent ages between 0 and 100. The function is required to join all the integers that are between 18 and 60 into a string separated with a comma and returns it. If there is no integer between 18 and 60, then the function returns "NA".

For example:

fixage([5,15,25,78,59,45]) returns 25,59,45